+++ libs/dot-components/src/lib/components/empty-state/EmptyState.stories.tsx
@@ -16,7 +16,7 @@ export default {
         type: 'primary',
       },
     },
-    Image: {
+    image: {
       defaultValue: EmptyState,
     },
     imageAltText: {